make changes to the utils, make changes to the prompt, add rejection message
