+++ ee/hogai/graph/root/prompts.py
@@ -63,6 +63,7 @@ The tool only retrieves a single query per call. If the user asks for multiple i
 
 Follow these guidelines when retrieving data:
 - If the same insight is already in the conversation history, reuse the retrieved data only when this does not violate the <data_analysis_guidelines> section (i.e. only when a presence-check, count, or sort on existing columns is enough).
+- If the previous analysis produced no data allow for query changes and start a new analysis with it.
 - If analysis results have been provided, use them to answer the user's question. The user can already see the analysis results as a chart - you don't need to repeat the table with results nor explain each data point.
 - If the retrieved data and any data earlier in the conversations allow for conclusions, answer the user's question and provide actionable feedback.
 - If there is a potential data issue, retrieve a different new analysis instead of giving a subpar summary. Note: empty data is NOT a potential data issue.
@@ -105,6 +106,32 @@ Follow these guidelines when searching insights:
 """.strip()
 )
 
+
+ROOT_INSIGHT_EDITING_PROMPT = """
+You are editing an existing insight. Your task is to describe the changes needed to update the current insight based on the user's latest request.
+
+## Key Rules:
+1. **Current insight schema is the source of truth** - only include fields that exist in the current insight
+2. **Ignore previous conversation changes** - if a user rejected a previous change in the UI, that change is no longer relevant
+3. **Focus on the user's most recent request** - describe only what they're asking for now
+4. **Include all current insight details** - preserve existing settings unless the user wants to change them
+
+## Process:
+1. Examine the current insight schema to understand what fields exist
+2. Identify the user's most recent request
+3. Create a description that combines current insight state + user's new request
+4. Exclude any rejected/removed changes from previous conversation
+
+## Example:
+- Current insight: tracks pageviews, no breakdown in its JSONdefinition
+- During the conversation: user asked for country breakdown, then removed it in UI (you do not see this UI removal, but you can assume it happened because that breakdown was not in the current insight schema)
+- Another user request: "add breakdown by exception"
+- Correct response: "Add breakdown by exception to the existing pageview insight that tracks pageviews"
+- Wrong response: "Add breakdown by exception and country to the existing pageview insight that tracks pageviews" (country was rejected)
+
+## Output:
+Provide a clear description of the insight with the user's requested changes applied. Mention all the existing fields in the insight + the user's new request. The tool will not have information about the insight.
+""".strip()
 
 ROOT_INSIGHT_DESCRIPTION_PROMPT = """
 Pick the most suitable visualization type for the user's question.
@@ -207,8 +234,13 @@ Description: {{.}}
 
 ROOT_INSIGHTS_CONTEXT_PROMPT = """
 # Insights
-The user has provided the following insights, which may be relevant to the question at hand:
+The user has provided the following insights, which is relevant to the question at hand:
 {{{insights}}}
+
+The existing insight needs to be updated to include the new request from the user. You should NOT delete any fields from the existing insight. You should only add or modify fields.
+The user can change their mind and remove a previous change in the UI of the app. This makes the previous changes not relevant to the current change request.
+It is your source of truth. Use it as the baseline for new edits.
+
 """.strip()
 
 ROOT_INSIGHT_CONTEXT_PROMPT = """
