+++ products/product_analytics/backend/max_tools.py
@@ -4,7 +4,7 @@ from collections.abc import AsyncIterator
 from langgraph.config import get_stream_writer
 from pydantic import BaseModel, Field
 
-from ee.hogai.graph.root.prompts import ROOT_INSIGHT_DESCRIPTION_PROMPT
+from ee.hogai.graph.root.prompts import ROOT_INSIGHT_DESCRIPTION_PROMPT, ROOT_INSIGHT_EDITING_PROMPT
 from ee.hogai.tool import MaxTool
 from ee.hogai.utils.types import AssistantState
 from posthog.schema import (
@@ -25,19 +25,25 @@ class EditCurrentInsightArgs(BaseModel):
     Edits the insight visualization the user is currently working on, by creating a query or iterating on a previous query.
     """
 
-    query_description: str = Field(
-        description="The new query to edit the current insight. Must include all details from the current insight plus any change on top of them. Include any relevant information from the current conversation, as the tool does not have access to the conversation."
-    )
+    query_description: str = Field(description="The new query to edit the current insight.")
     query_kind: str = Field(description=ROOT_INSIGHT_DESCRIPTION_PROMPT)
 
 
 class EditCurrentInsightTool(MaxTool):
     name: str = "create_and_query_insight"  # this is the same name as the "insights" tool in the backend, as this overwrites that tool's functionality to be able to send the result to the frontend
-    description: str = (
-        "Update the insight the user is currently working on, based on the current insight's JSON schema."
-    )
+    description: str = ROOT_INSIGHT_EDITING_PROMPT
     thinking_message: str = "Editing your insight"
-    root_system_prompt_template: str = "The user is currently editing an insight (aka query). Here is that insight's current definition, which can be edited using the `create_and_query_insight` tool:\n```json\n{current_query}\n```"
+    root_system_prompt_template: str = """The user is currently editing an insight (aka query). Here is that insight's current definition, which can be edited using the `create_and_query_insight` tool:
+
+```json
+{current_query}
+```
+
+IMPORTANT: YOU MUST ALWAYS describe the most recent state of the insight to the tool and the most recent change the user has requested.
+THE CURRENT INSIGHT DEFINITION IS YOUR SOURCE OF TRUTH.
+
+""".strip()
+
     args_schema: type[BaseModel] = EditCurrentInsightArgs
 
     async def _arun_impl(self, query_kind: str, query_description: str) -> tuple[str, None]:
@@ -45,6 +51,10 @@ class EditCurrentInsightTool(MaxTool):
 
         if "current_query" not in self.context:
             raise ValueError("Context `current_query` is required for the `create_and_query_insight` tool")
+
+        # Fix the type errors by ensuring team and user are not None
+        if self._team is None or self._user is None:
+            raise ValueError("Team and user must be available for the `create_and_query_insight` tool")
 
         graph = InsightsAssistantGraph(self._team, self._user).compile_full_graph()
         state = self._state
